refactor(service): Move the service firebase-cards in an other folder
