+++ angular_front/src/app/services/firebase-cards/firebase-cards.service.ts
